requirements.txt
scapy
flask
flask-socketio
python-socketio
eventlet
requests
python-dotenv
